Add support for encrypted files
I don't like exposing some part of my config. Even though I use e.g.
PassCmd in mbsyncrc, and don't include my password there, if I commit it
as is, I'd be exposing my username, ports and authentication mechanisms.
There might be security vulnerabilities which would be immediately
visible and so on. It's not like these things are top secret, they
should be somewhat easily discoverable, but if it is within my reach,
I'd like to not expose it for free.